leptos 0.9; fix macos file system daemon watching config file logic

+++ desktop/close_ports.sh
@@ -3,4 +3,6 @@ SYRE_SERVER_PROCESS_NAME="trunk"
 SYRE_SERVER_PORT=1420
 
 pid=$(lsof -i :$SYRE_SERVER_PORT | grep $SYRE_SERVER_PROCESS_NAME | cut -f 2 -w)
-kill $pid
+if [ -n "$pid" ]; then
+	kill $pid
+fi
